Stacked Borrows: test raw-ref-to-field with raw ptr tracking
